added a base class for nodeElem constraints that is called by constraintWarehouse and nonlinearSystemBase. This builds and all of the tests pass.
