+++ framework/src/constraints/NodeElemConstraint.C
@@ -13,34 +13,22 @@
 #include "Assembly.h"
 #include "MooseEnum.h"
 #include "MooseMesh.h"
-#include "MooseVariableFE.h"
-#include "SystemBase.h"
 
 #include "libmesh/string_to_enum.h"
 
 InputParameters
 NodeElemConstraint::validParams()
 {
-  InputParameters params = Constraint::validParams();
+  InputParameters params = NodeElemConstraintBase::validParams();
   params.addRequiredParam<SubdomainName>("secondary", "secondary block id");
   params.addRequiredParam<SubdomainName>("primary", "primary block id");
-  params.addRequiredCoupledVar("primary_variable",
-                               "The variable on the primary side of the domain");
-
   return params;
 }
 
 NodeElemConstraint::NodeElemConstraint(const InputParameters & parameters)
-  : Constraint(parameters),
-    // The secondary side is at nodes (hence passing 'true').  The neighbor side is the primary side
-    // and it is not at nodes (so passing false)
-    NeighborCoupleableMooseVariableDependencyIntermediateInterface(this, true, false),
-    NeighborMooseVariableInterface<Real>(
-        this, true, Moose::VarKindType::VAR_SOLVER, Moose::VarFieldType::VAR_FIELD_STANDARD),
-
+  : NodeElemConstraintBase(parameters),
     _secondary(_mesh.getSubdomainID(getParam<SubdomainName>("secondary"))),
     _primary(_mesh.getSubdomainID(getParam<SubdomainName>("primary"))),
-    _var(_sys.getFieldVariable<Real>(_tid, parameters.get<NonlinearVariableName>("variable"))),
 
     _primary_q_point(_assembly.qPoints()),
     _primary_qrule(_assembly.qRule()),
@@ -53,7 +41,6 @@ NodeElemConstraint::NodeElemConstraint(const InputParameters & parameters)
     _phi_secondary(1),
     _test_secondary(1), // One entry
 
-    _primary_var(*getVar("primary_variable", 0)),
     _primary_var_num(_primary_var.number()),
 
     _phi_primary(_assembly.phiNeighbor(_primary_var)),
@@ -67,13 +54,11 @@ NodeElemConstraint::NodeElemConstraint(const InputParameters & parameters)
     _grad_u_primary(_primary_var.gradSlnNeighbor()),
 
     _dof_map(_sys.dofMap()),
-    _node_to_elem_map(_mesh.nodeToElemMap()),
+    _node_to_elem_map(_mesh.nodeToElemMap())
 
-    _overwrite_secondary_residual(false)
 {
   _mesh.errorIfDistributedMesh("NodeElemConstraint");
 
-  addMooseVariableDependency(&_var);
   // Put a "1" into test_secondary
   // will always only have one entry that is 1
   _test_secondary[0].push_back(1);
@@ -83,14 +68,6 @@ NodeElemConstraint::~NodeElemConstraint()
 {
   _phi_secondary.release();
   _test_secondary.release();
-}
-
-void
-NodeElemConstraint::computeSecondaryValue(NumericVector<Number> & current_solution)
-{
-  const dof_id_type & dof_idx = _var.nodalDofIndex();
-  _qp = 0;
-  current_solution.set(dof_idx, computeQpSecondaryValue());
 }
 
 void
@@ -218,10 +195,4 @@ NodeElemConstraint::getConnectedDofIndices(unsigned int var_num)
     else
       _phi_secondary[j][_qp] = 0.0;
   }
-}
-
-bool
-NodeElemConstraint::overwriteSecondaryResidual()
-{
-  return _overwrite_secondary_residual;
 }
